+++ src/main/resources/i18n/messages.properties
@@ -991,7 +991,7 @@ project.settings.coverage.maximum_coverage=Maximum Coverage:
 # ========================================================================================== #
 # PROJECT METADATA TEMPLATES                                                                 #
 # ========================================================================================== #
-ProjectMetadataTemplates.title=Metadata Templates
+ProjectMetadataTemplates.title=Templates
 ProjectMetadataTemplates.fields={0} fields
 
 MetadataFields.title=Metadata Fields
