+++ src/main/webapp/resources/js/pages/projects/settings/components/SettingsNav.jsx
@@ -23,14 +23,9 @@ export default function SettingsNav({ path }) {
       <Menu.Item key="groups">
         <Link to="groups">{i18n("project.settings.page.groups")}</Link>
       </Menu.Item>
-      <Menu.Item key="metadata-fields">
-        <Link className="t-m-field-link" to="metadata-fields">
+      <Menu.Item key="metadata">
+        <Link className="t-m-field-link" to="metadata/fields">
           {i18n("MetadataFields.title")}
-        </Link>
-      </Menu.Item>
-      <Menu.Item key="metadata-templates">
-        <Link className="t-m-template-link" to="metadata-templates">
-          {i18n("ProjectMetadataTemplates.title")}
         </Link>
       </Menu.Item>
     </Menu>
