+++ src/test/java/ca/corefacility/bioinformatics/irida/ria/integration/pages/projects/ProjectMetadataPage.java
@@ -43,7 +43,7 @@ public class ProjectMetadataPage extends AbstractPage {
 	}
 
 	public static ProjectMetadataPage goTo(WebDriver driver) {
-		get(driver, "/projects/1/settings/metadata-fields");
+		get(driver, "/projects/1/settings/metadata/fields");
 		return PageFactory.initElements(driver, ProjectMetadataPage.class);
 	}
 
@@ -89,7 +89,7 @@ public class ProjectMetadataPage extends AbstractPage {
 				.sendKeys(description);
 		createTemplateModal.findElement(By.className("t-create-modal-ok"))
 				.click();
-		wait.until(ExpectedConditions.urlContains("/metadata-templates"));
+		wait.until(ExpectedConditions.urlContains("/metadata/templates"));
 	}
 
 	public void gotoTemplate(String name) {
@@ -99,7 +99,7 @@ public class ProjectMetadataPage extends AbstractPage {
 			if (element.getText()
 					.equalsIgnoreCase(name)) {
 				element.click();
-				wait.until(ExpectedConditions.urlContains("/metadata-templates/"));
+				wait.until(ExpectedConditions.urlContains("/metadata/templates/"));
 				break;
 			}
 		}
