+++ src/test/java/ca/corefacility/bioinformatics/irida/ria/integration/projects/ProjectMetadataIT.java
@@ -40,7 +40,7 @@ public class ProjectMetadataIT extends AbstractIridaUIITChromeDriver {
 		page.selectMetadataField("Symptoms");
 		page.selectMetadataField("Exposures");
 		page.createNewTemplate("Special Template", "Long description");
-		Assert.assertTrue("Should be on a template specific page", driver().getCurrentUrl().matches("(.*)/metadata-templates/\\d+"));
+		Assert.assertTrue("Should be on a template specific page", driver().getCurrentUrl().matches("(.*)/metadata/templates/\\d+"));
 		final String newTemplateName = "An awesome name";
 		final String currentName = page.getTemplateName();
 		page.editTemplateName(newTemplateName);
